Resolves a node to the first resolvable parent
The goal is to get the symbol under the cursor if possible but to
fallback to the parent symbol if the one under the cusor is not
resolvable.

So if the cursor is inside a method block but on an empty line, it will
be resolve to the method symbol.
Same if the cursor is between two class members, then the symbol should
be the one of the class.