DeclaredMemberTypeResolver handle `iterable` and `resource`
Currently the tolerant parser does not identifies these types as `Token`
like it does for `callable` but as `QualifiedName`.
The problem is that we create a type which a classname for them so when
we try to import the memthods of an interface that uses one of these
types, the code generator also include a use statement for them.

This is my attempt to fix that, it seems to work for now, I have updated
the tests I found relevant and we will see if I didn't forgot anything.